+++ variables.py
@@ -1,0 +1,2 @@
+UserOfJobId = {}
+qstatqLst = []
